requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-mistralai>=0.1.0
langchain-pinecone>=0.1.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pandas>=2.0.0
streamlit>=1.32.0
requests>=2.31.0
transformers>=4.30.0
torch>=2.0.0
jq>=1.6.0  # Required for JSON parsing with jq expressions
supabase>=2.0.0
atlassian-python-api>=3.41.0
nltk>=3.8.1
